+++ src/lockfile.ts
@@ -12,7 +12,8 @@ export type LockFileConfigV0 = {
 export type LockFilePackageEntry = {
   version?: string
   file?: boolean,
-  replaced?: string
+  replaced?: string,
+  signature?: string
 }
 
 export type LockFileConfigV1 = {
@@ -83,10 +84,11 @@ export const addPackageToLockfile = (
   packages: ({ name: string } & LockFilePackageEntry)[],
   options: { workingDir: string }) => {
   const lockfile = readLockfile(options)
-  packages.forEach(({ name, version, file, replaced }) => {
+  packages.forEach(({ name, version, file, replaced, signature }) => {
     let old = lockfile.packages[name] || {}
     lockfile.packages[name] = {}
     version && (lockfile.packages[name].version = version)
+    signature && (lockfile.packages[name].signature = signature)
     file && (lockfile.packages[name].file = true)
     if (replaced || old.replaced) {
       lockfile.packages[name].replaced = replaced || old.replaced